Apply clippy fixes, test cfg-gating, and code quality improvements across 159 modules
Cleans up accumulated working-tree changes from the automation session:
- Remove unnecessary 'as u32' casts flagged by clippy
- Gate C extern declarations behind #[cfg(not(test))] to fix test compilation
- Replace manual loop indexing with iterator-based patterns
- Add missing safety doc comments for unsafe extern fns
- Update .llxprt/LLXPRT.md with current project status
- Add sdl_capture.rs (ABI-authoritative SDL surface accessors)
- Add proptest-regressions/ directory

All quality gates pass: fmt, clippy, 3110 lib tests.

+++ rust/src/battle/ai.rs
@@ -310,11 +310,11 @@ mod tests {
         };
         let flags = input.to_status_flags();
         use super::super::ship_runtime::{LEFT, RIGHT, SPECIAL, THRUST, WEAPON};
-        assert!(flags & (LEFT as u32) != 0);
-        assert!(flags & (THRUST as u32) != 0);
-        assert!(flags & (WEAPON as u32) != 0);
-        assert!(flags & (RIGHT as u32) == 0);
-        assert!(flags & (SPECIAL as u32) == 0);
+        assert!(flags & LEFT != 0);
+        assert!(flags & THRUST != 0);
+        assert!(flags & WEAPON != 0);
+        assert!(flags & RIGHT == 0);
+        assert!(flags & SPECIAL == 0);
     }
 
     #[test]
